Reorganize WASM build
